+++ .eslintrc.js
@@ -462,7 +462,7 @@ module.exports = {
       },
     },
     {
-      files: ['packages/graphiql-react/**'],
+      files: ['packages/{graphiql-react,graphiql}/**'],
       plugins: ['react-compiler'],
       rules: {
         '@typescript-eslint/no-restricted-imports': [