+++ packages/graphql-language-service-server/src/__tests__/MessageProcessor.spec.ts
@@ -28,7 +28,7 @@ jest.mock('@whatwg-node/fetch', () => {
   };
 });
 
-const mockSchema = (schema: GraphQLSchema) => {
+function mockSchema(schema: GraphQLSchema) {
   const introspectionResult = {
     data: introspectionFromSchema(schema, {
       descriptions: true,
@@ -43,7 +43,7 @@ const mockSchema = (schema: GraphQLSchema) => {
       body: introspectionResult,
     },
   });
-};
+}
 
 const defaultFiles = [
   ['query.graphql', 'query { bar ...B }'],